+++ pkg/querier/tripperware/roundtrip.go
@@ -143,7 +143,7 @@ func NewQueryTripperware(
 				tenantIDs, err := tenant.TenantIDs(r.Context())
 				// This should never happen anyways because we have auth middleware before this.
 				if err != nil {
-					return nil, err
+					return nil, httpgrpc.Errorf(http.StatusBadRequest, err.Error())
 				}
 				now := time.Now()
 				userStr := tenant.JoinTenantIDs(tenantIDs)
@@ -162,8 +162,7 @@ func NewQueryTripperware(
 
 					expr, err := parser.ParseExpr(query)
 					if err != nil {
-						// If query is invalid, no need to go through tripperwares for further splitting.
-						return next.RoundTrip(r)
+						return nil, httpgrpc.Errorf(http.StatusBadRequest, err.Error())
 					}
 
 					reqStats := stats.FromContext(r.Context())
@@ -189,7 +188,10 @@ func NewQueryTripperware(
 						return next.RoundTrip(r)
 					}
 					analysis, err := queryAnalyzer.Analyze(query)
-					if err != nil || !analysis.IsShardable() {
+					if err != nil {
+						return nil, httpgrpc.Errorf(http.StatusBadRequest, err.Error())
+					}
+					if !analysis.IsShardable() {
 						return next.RoundTrip(r)
 					}
 					return instantQuery.RoundTrip(r)